Refactor Mininet cluster
1. Create cluster/{net, link, node, placer, cleanup}
2. Rename clustercli to cluster/cli
3. Append package name in setup.py
