+++ examples/clusterSanity.py
@@ -4,9 +4,9 @@
 A sanity check for cluster edition
 '''
 
-from mininet.examples.cluster import MininetCluster
+from mininet.cluster.net import MininetCluster
 from mininet.log import setLogLevel
-from mininet.examples.clustercli import ClusterCLI as CLI
+from mininet.cluster.cli import ClusterCLI as CLI
 from mininet.topo import SingleSwitchTopo
 
 def clusterSanity():
